requirements.txt
openai>=1.3.0
beautifulsoup4>=4.12.0
markdownify>=0.11.6
PyYAML>=6.0
requests>=2.31.0
tiktoken>=0.5.0
pypdf>=3.17.0
python-dotenv>=1.0.0